refactor: fix naming convention in card component
