+++ apps/web/src/modules/job/components/card/components/footer/styled.ts
