+++ apps/web/src/modules/job/components/card/components/footer/type.ts
@@ -1,0 +1,6 @@
+export interface FooterProps {
+  gender: string
+  actorCount: number
+  wage: number
+  dueDate: Date
+}
