+++ apps/web/src/modules/job/components/card/components/header/styled.ts
