+++ apps/web/src/modules/job/components/card/components/header/type.ts
@@ -1,0 +1,5 @@
+export interface HeaderProps {
+  title: string
+  companyName: string
+  castingImage: string
+}
