+++ apps/web/src/modules/job/components/card/components/shootingDetail/styled.ts
